+++ server/apps/publish/__init__.py
@@ -12,10 +12,8 @@
 import logging
 
 import superdesk
-from apps.publish.archive_publish import ArchivePublishResource, ArchivePublishService, \
+from apps.publish.content import ArchivePublishResource, ArchivePublishService, \
     KillPublishResource, KillPublishService, CorrectPublishResource, CorrectPublishService
-from apps.publish.subscribers import SubscribersResource, SubscribersService
-from apps.publish.publish_queue import PublishQueueResource, PublishQueueService
 from apps.publish.published_item import PublishedItemResource, PublishedItemService
 from apps.publish.commands import RemoveExpiredPublishContent  # noqa
 from superdesk.celery_app import celery
@@ -37,14 +35,6 @@ def init_app(app):
     endpoint_name = 'archive_correct'
     service = CorrectPublishService(endpoint_name, backend=get_backend())
     CorrectPublishResource(endpoint_name, app=app, service=service)
-
-    endpoint_name = 'subscribers'
-    service = SubscribersService(endpoint_name, backend=get_backend())
-    SubscribersResource(endpoint_name, app=app, service=service)
-
-    endpoint_name = 'publish_queue'
-    service = PublishQueueService(endpoint_name, backend=get_backend())
-    PublishQueueResource(endpoint_name, app=app, service=service)
 
     endpoint_name = 'published'
     service = PublishedItemService(endpoint_name, backend=get_backend())
