+++ server/apps/publish/formatters/anpa_formatter.py
@@ -8,7 +8,7 @@
 # AUTHORS and LICENSE files distributed with this source code, or
 # at https://www.sourcefabric.org/superdesk/license
 
-from apps.publish.formatters import Formatter
+from superdesk.publish.formatters import Formatter
 from apps.publish.formatters.aap_formatter_common import map_priority
 import superdesk
 from superdesk.errors import FormatterError
