--- server/apps/vocabularies/__init__.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8; -*-
-#
-# This file is part of Superdesk.
-#
-# Copyright 2013, 2014 Sourcefabric z.u. and contributors.
-#
-# For the full copyright and license information, please see the
-# AUTHORS and LICENSE files distributed with this source code, or
-# at https://www.sourcefabric.org/superdesk/license
-
-from .vocabularies import VocabulariesResource, VocabulariesService
-from .command import VocabulariesPopulateCommand  # noqa
-import superdesk
-
-
-def init_app(app):
-    endpoint_name = 'vocabularies'
-    service = VocabulariesService(endpoint_name, backend=superdesk.get_backend())
-    VocabulariesResource(endpoint_name, app=app, service=service)
